% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_velocity.R
\name{compute_velocity_table}
\alias{compute_velocity_table}
\title{Compute velocity for x, y, and its amplited}
\usage{
compute_velocity_table(x, y, trial, sample_rate, velocity_time_window)
}
\arguments{
\item{x}{vector with x coordinates in \emph{degrees of visual angle}}

\item{y}{vector with y coordinates in \emph{degrees of visual angle}}

\item{trial}{vector with trial index}

\item{sample_rate}{sample rate in Hz}

\item{velocity_time_window}{Time window for velocity computation in \emph{milliseconds}}
}
\value{
\code{data.frame} with columns \code{x}, \code{y}, and \code{amp}
}
\description{
Compute velocity for x, y, and its amplited
}
\examples{
compute_velocity_df(rnorm(1000), rnorm(1000), rep(1, 1000), 250, 20)
}
