% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_velocity_via_ek}
\alias{compute_velocity_via_ek}
\title{Compute velocity via Engbert & Kliegl (2003) algorithm.}
\usage{
compute_velocity_via_ek(x, trial, time_window_in_samples, delta_t)
}
\arguments{
\item{x}{Vector of coordinates}

\item{trial}{Vector with trial labels, so that velocity is computed only within trials.}

\item{time_window_in_samples}{Width of window for velocity computation in samples.}

\item{delta_t}{Duration of a single frame (1 / sampling rate).}
}
\value{
Velocity vector
}
\description{
Compute velocity via Engbert & Kliegl (2003) algorithm.
See formula #1 in the manuscript.
}
