% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_microsaccades.R
\name{extract_microsaccades}
\alias{extract_microsaccades}
\title{Extracts microsaccades from samples using one of the implemented algorithms.}
\usage{
extract_microsaccades(
  x,
  y,
  sample_rate,
  method = "ek",
  binocular = "merge",
  trial = NULL,
  options = list()
)
}
\arguments{
\item{x}{Horizontal coordinate, ether a vector for monocular data or a two-column matrix for binocular data.}

\item{y}{Vertical coordinate, ether a vector for monocular data or a two-column matrix for binocular data.}

\item{sample_rate}{Sampling rate in Hz.}

\item{method}{A string with a method name. Currently implemented are Engbret & Kliegl (2003) (\code{"ek"}, default).}

\item{binocular}{Specifies how a binocular data is treated. Options are \code{"cyclopean"} (binocular data is
converted to an average cyclopean image before microsaccades are extracted), \code{"independent"} (microsaccades
are extracted independently for each eye), \code{"merge"} (default, microsaccades are extracted for each eye
independently but microsaccades from different eyes that temporally overlap are averaged into a binocular
microsaccade).}

\item{trial}{Optional vector with trial ID. If omitted, all samples are assumed to belong to a single trial.}

\item{options}{A named list with options for a specific method.}
}
\value{

}
\description{
Extracts microsaccades from samples using one of the implemented algorithms.
}
