% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_nh.R
\name{extract_ms_nh}
\alias{extract_ms_nh}
\title{Extract microsaccades using an algorithm proposed by Nyström and Holmqvist (2010) \doi{10.3758/BRM.42.1.188}.}
\usage{
extract_ms_nh(
  x,
  y,
  vel,
  acc,
  sample_rate,
  trial,
  sg_filter_order = 2,
  max_velocity = 1000,
  max_acceleration = 1e+05,
  initial_velocity_threshold = 100
)
}
\arguments{
\item{x}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{y}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{vel}{Velocity data.frame with columns \code{x}, \code{y}, \code{amp}.}

\item{acc}{Acceleration data.frame with columns \code{x}, \code{y}, \code{amp}.}

\item{sample_rate}{Sample rate in Hz.}

\item{trial}{Trial id, so that trial borders are respected when computing velocity and saccades.}

\item{sg_filter_order}{Order of Savitzky-Golay filter. Please refer to Nyström and Holmqvist (2010) for details.}

\item{max_velocity}{Maximal physiologically plausible velocity in °/s. Please refer to Nyström and Holmqvist (2010) for details.}

\item{max_acceleration}{Maximal physiologically plausible acceleration in °/s*s. Please refer to Nyström and Holmqvist (2010) for details.}
}
\value{
logical vector marking samples that belong to saccades
}
\description{
Extract microsaccades using an algorithm proposed by Nyström and Holmqvist (2010) \doi{10.3758/BRM.42.1.188}.
}
\examples{
# Do not run this function directly, use vote_on_samples() or extract_saccades()
}
\seealso{
\code{\link{vote_on_samples}}, \code{\link{extract_saccades}}
}
