% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_nh.R
\name{filter_via_savitzky_golay}
\alias{filter_via_savitzky_golay}
\title{Smooths signal using Savitzky-Golay and then shifts the filtered signal back}
\usage{
filter_via_savitzky_golay(x, sg_order)
}
\arguments{
\item{x}{vector of float}

\item{sgOrder}{integer, order of the filter}
}
\value{
vector of float
}
\description{
Smooths signal using Savitzky-Golay and then shifts the filtered signal back
}
\examples{
filter_via_savitzky_golay(rnorm(1000), 2)
}
\keyword{internal}
