% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_matrix.R
\name{input_to_matrix}
\alias{input_to_matrix}
\title{Converts inputs to matrix}
\usage{
input_to_matrix(x)
}
\arguments{
\item{x}{vector, matrix, data.frame/tibble}
}
\value{
matrix
}
\description{
Converts inputs (vector, matrix, data.frame/tibble) to 
a matrix preserving number of columns
}
\examples{
input_to_matrix(1:5)
}
\keyword{internal}
