% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monocular_ten_trials}
\alias{monocular_ten_trials}
\title{A monocular multi-trial recording}
\format{
A data frame with 14353 rows and 4 variables:
\describe{
  \item{trial}{Trial index.}
  \item{x}{X coordinate in degrees of visual angle.}
  \item{y}{Y coordinate in degrees of visual angle.}
  \item{time}{Sample time in milliseconds.}
}
}
\usage{
monocular_ten_trials
}
\description{
A monocular recording, 10 trials, sampling rate 500 Hz.
}
\keyword{datasets}
