% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_or_default.R
\name{option_or_default}
\alias{option_or_default}
\title{Extract value for a named list or use default if key is missing}
\usage{
option_or_default(options, key, default)
}
\arguments{
\item{options}{Named list}

\item{key}{String key}

\item{default}{Default value to be returned, if key is missing.}
}
\value{
Value from a list or default value
}
\description{
Extract value for a named list or use default if key is missing
}
\examples{
option_or_default(list("A"), "A", 20)
option_or_default(list("A"), "B", 20)
}
