% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccadr.R
\docType{package}
\name{saccadr-package}
\alias{saccadr}
\alias{saccadr-package}
\title{saccadr: Extract Saccades via an Ensemble of Methods Approach}
\description{
Extract (Micro)Saccades From Gaze Samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexander-pastukhov/saccadr/}
  \item \url{https://alexander-pastukhov.github.io/saccadr/}
  \item Report bugs at \url{https://github.com/alexander-pastukhov/saccadr/issues/}
}

}
\author{
\strong{Maintainer}: Alexander Pastukhov \email{pastukhov.alexander@gmail.com} (\href{https://orcid.org/0000-0002-8738-8591}{ORCID})

}
\keyword{internal}
