% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_trial}
\alias{single_trial}
\title{A single trial monocular samples recorded at 500 Hz.}
\format{
A data frame with 1006 rows and 2 variables:
\describe{
  \item{x}{X coordinate in degrees of visual angle.}
  \item{y}{Y coordinate in degrees of visual angle.}
}
}
\usage{
single_trial
}
\description{
A single trial monocular samples recorded at 500 Hz.
}
\keyword{datasets}
