% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_trial_binocular}
\alias{single_trial_binocular}
\title{A single trial binocular recording.}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
  \item{trial}{Trial index.}
  \item{time_rel}{Sample time in milliseconds relative to the trial start.}
  \item{xL}{X coordinate for the left eye in degrees of visual angle.}
  \item{xR}{X coordinate for the right eye in degrees of visual angle.}
  \item{yL}{Y coordinate for the left eye in degrees of visual angle.}
  \item{yR}{Y coordinate for the right eye in degrees of visual angle.}
}
}
\usage{
single_trial_binocular
}
\description{
A single trial binocular recording sampled at 1000 Hz.
}
\keyword{datasets}
