% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_on_samples.R
\name{vote_on_samples}
\alias{vote_on_samples}
\title{Computes ensemble vote for each sample using selected methods.}
\usage{
vote_on_samples(
  x,
  y,
  sample_rate,
  trial = NULL,
  methods = list("ek", "om", "nh"),
  options = NULL,
  velocity_time_window = 20,
  binocular = "merge",
  normalize = TRUE
)
}
\arguments{
\item{x}{Horizontal coordinate, either a vector for monocular data or a two-column matrix for binocular data.}

\item{y}{Vertical coordinate, either a vector for monocular data or a two-column matrix for binocular data.}

\item{sample_rate}{Sampling rate in Hz.}

\item{trial}{Optional vector with trial ID. If omitted, all samples are assumed to belong to a single trial.}

\item{methods}{A \emph{list} (not a vector!) with names of package methods (character) or external functions that
implement sample classification (see vignette on using custom methods). Package methods include
Engbret & Kliegl (2003) (\code{"ek"}), Otero-Millan et al. (\code{"om"}), Nyström and Holmqvist (2010) (\code{"nh"}).}

\item{options}{A named list with options, see documentation on specific method for details.}

\item{velocity_time_window}{Time window for computing velocity and acceleration in milliseconds.}

\item{binocular}{Specifies how a binocular data is treated. Options are \code{"cyclopean"} (binocular data is
converted to an average cyclopean image before saccades are extracted), \code{"monocular"} (saccades
are extracted independently for each eye), \code{"merge"} (default, saccades are extracted for each eye
independently but saccades from different eyes that temporally overlap are averaged into a binocular
saccade). Note that \code{binocular = "merge"} is overridden by \code{normalize = FALSE},
votes are left as they are per method and eye.}

\item{normalize}{Logical, whether to aggregate and normalize votes to 0..1 range. Note that
\code{normalize = FALSE} overrides \code{binocular = "merge"}, votes are left as they are per method and eye.}
}
\value{
list Either a list of matrices (one per eye) with votes for each method (\code{normalized = FALSE}) or
a matrix with normalized (0..1) vote on each sample either for each eye or for single cyclopean or merged
binocular time series (\code{normalized = TRUE}).
}
\description{
This function is called by \code{\link{extract_saccades}} but can be used directly,
if you are interested in the vote-per-sample.
}
\examples{
data(single_trial)
vote <- vote_on_samples(x = single_trial[['x']],
                        y = single_trial[['y']],
                        methods = list("ek", "om", "nh"),
                        sample_rate = 500,
                        normalize = FALSE)
}
\seealso{
\code{\link{extract_saccades}}, \code{\link{method_ek}}, \code{\link{method_om}}, \code{\link{method_nh}}
}
