% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_method.R
\name{extract_ms_ek}
\alias{extract_ms_ek}
\title{Extract microsaccades using an algorithm proposed by Engbert and Kliegl (2003) \doi{10.1016/S0042-6989(03)00084-1}}
\usage{
extract_ms_ek(
  x,
  y,
  vel,
  acc,
  sample_rate,
  trial,
  velocity_threshold = 6,
  sd_fun = sd_via_median_estimator,
  minimal_duration_ms = 12,
  minimal_separation_ms = 12
)
}
\arguments{
\item{x}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{y}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{vel}{Velocity data.frame with columns \code{x}, \code{y}, \code{amp}.}

\item{acc}{Acceleration data.frame with columns \code{x}, \code{y}, \code{amp}.}

\item{sample_rate}{Sample rate in Hz.}

\item{trial}{Trial id, so that trial borders are respected when computing velocity and saccades.}

\item{velocity_threshold}{Velocity threshold for microsaccade detection in medians. Please refer to Engbert and Kliegl (2003) for details.}

\item{sd_fun}{Function used to compute standard deviation for velocities. Defaults to \code{sd_via_median_estimator}, as per formula #2 in Engbert and Kliegl (2003). Can be replaced with \code{mad}, \code{sd}, etc.}

\item{minimal_duration_ms}{Minimal duration of a microsaccade in milliseconds. Please refer to Engbert and Kliegl (2003) for details.}

\item{minimal_separation_ms}{A minimal required time gap between two microsaccades. Please refer to Engbert and Kliegl (2003) for details.}
}
\value{
logical vector marking samples that belong to saccades
}
\description{
Extract microsaccades using an algorithm proposed by Engbert and Kliegl (2003) \doi{10.1016/S0042-6989(03)00084-1}
}
\examples{
# Do not run this function directly, use vote_on_samples() or extract_saccades()
}
\seealso{
\code{\link{vote_on_samples}}, \code{\link{extract_saccades}}
}
