% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ek_method.R
\name{extract_ms_ek}
\alias{extract_ms_ek}
\title{Extract microsaccades using an algorithm proposed by Engbert and Kliegl (2003)}
\usage{
extract_ms_ek(
  x,
  y,
  sample_rate,
  trial = NULL,
  velocity_time_window = 20,
  velocity_threshold = 6,
  sd_fun = sd_via_median_estimator,
  minimal_duration_ms = 12,
  minimal_separation_ms = 0
)
}
\arguments{
\item{x}{Gaze x coordinate, arbitrary units as threshold velocity is computed in units of standard deviation.}

\item{y}{Gaze x coordinate, arbitrary units as threshold velocity is computed in units of standard deviation.}

\item{sample_rate}{Sample rate of the recording in Hz.}

\item{trial}{Trial id. Trial borders are respected when computing velocity.
All samples are assumed to belong to the same trial, if \code{trial = NULL}.}

\item{velocity_time_window}{Time span in milliseconds relative to the sample
that is used  to compute the velocity. Defaults to 20 ms.}

\item{velocity_threshold}{Velocity threshold for microsaccade detection
in medians. Defaults to 6.}

\item{sd_fun}{Function used to compute standard deviation for velocities.
Defaults to \code{sd_via_median_estimator}, as per formula #2 in the paper.
Could be replaced with \code{mad}, \code{sd}, etc.}

\item{minimal_duration_ms}{Minimal duration of a microsaccade in milliseconds.
All shorter microsaccades are discarded. Defaults to 12 ms.}

\item{minimal_separation_ms}{A minimal required time gap between two microsaccades.
Two microsaccades that a separated by a smaller time gap are merged into a single
microsaccade. Defaults to 0 ms.}
}
\value{
\code{data.frame} with
}
\description{
Extract microsaccades using an algorithm proposed by Engbert and
Kliegl (2003).
}
