% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_nh.R
\name{extract_ms_nh}
\alias{extract_ms_nh}
\title{Extract microsaccades using an algorithm proposed by Nyström and Holmqvist (2010)}
\usage{
extract_ms_nh(
  x,
  y,
  vel,
  acc,
  sample_rate,
  trial,
  sg_filter_order = 2,
  max_velocity = 1000,
  max_acceleration = 1e+05,
  initial_velocity_threshold = 100,
  velocity_threshold = 6,
  sd_fun = sd_via_median_estimator,
  minimal_duration_ms = 12,
  minimal_separation_ms = 12
)
}
\description{
Extract microsaccades using an algorithm proposed by Nyström and Holmqvist (2010)
}
