% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_method.R
\name{extract_ms_om}
\alias{extract_ms_om}
\title{Extract microsaccades using an algorithm proposed by Otero-Millan et al. (2014)}
\usage{
extract_ms_om(
  x,
  y,
  vel,
  acc,
  sample_rate,
  trial,
  velocity_threshold = 6,
  sd_fun = sd_via_median_estimator,
  minimal_duration_ms = 12,
  minimal_separation_ms = 12
)
}
\arguments{
\item{x}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{y}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{vel}{Velocity data.frame with columns \code{x}, \code{y}, \code{amp}.}

\item{acc}{Acceleration data.frame with columns \code{x}, \code{y}, \code{amp}.}
}
\description{
Extract microsaccades using an algorithm proposed by Otero-Millan et al. (2014)
}
\examples{
# do not run this function directly, use extract_microsaccades
}
