% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_method.R
\name{extract_ms_om}
\alias{extract_ms_om}
\title{Extract microsaccades using an algorithm proposed by Otero-Millan et al. (2014)}
\usage{
extract_ms_om(
  x,
  y,
  vel,
  acc,
  sample_rate,
  trial,
  minimal_inter_peak_time_ms = 30,
  maximal_peaks_per_second = 5,
  velocity_threshold_deg_per_sec = 3,
  pca_variance_threshold = 0.05
)
}
\arguments{
\item{x}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{y}{Gaze x coordinate, _arbitrary units_ as threshold velocity is computed in units of standard deviation.}

\item{vel}{Velocity data.frame with columns \code{x}, \code{y}, \code{amp}.}

\item{acc}{Acceleration data.frame with columns \code{x}, \code{y}, \code{amp}.}
}
\description{
Extract microsaccades using an algorithm proposed by Otero-Millan et al. (2014)
}
\examples{
# do not run this function directly, use extract_microsaccades
}
