% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_saccades.R
\name{extract_saccades}
\alias{extract_saccades}
\title{Extract saccades from samples using votes from selected methods.}
\usage{
extract_saccades(
  x,
  y,
  sample_rate,
  trial = NULL,
  methods = list("ek", "om", "nh"),
  options = NULL,
  velocity_time_window = 20,
  binocular = "merge",
  vote_threshold = ((length(methods) - 1)/length(methods)) * 0.99,
  minimal_duration_ms = 12,
  minimal_separation_ms = 12,
  return_votes = FALSE
)
}
\arguments{
\item{x}{Horizontal coordinate, either a vector for monocular data or a two-column matrix for binocular data.}

\item{y}{Vertical coordinate, either a vector for monocular data or a two-column matrix for binocular data.}

\item{sample_rate}{Sampling rate in Hz.}

\item{trial}{Optional vector with trial ID. If omitted, all samples are assumed to belong to a single trial.}

\item{methods}{A \emph{list} (not a vector!) with names of package methods (character) or external functions that
implement sample classification (see vignette on using custom methods). Package methods include
Engbret & Kliegl (2003) (\code{"ek"}), Otero-Millan et al. (\code{"om"}), Nyström and Holmqvist (2010) (\code{"nh"}).}

\item{options}{A named list with options for a specific method, see documentation on specific method for details.}

\item{velocity_time_window}{Time window for computing velocity and acceleration in milliseconds.}

\item{binocular}{Specifies how a binocular data is treated. Options are \code{"cyclopean"} (binocular data is
converted to an average cyclopean image before saccades are extracted), \code{"monocular"} (saccades
are extracted independently for each eye), \code{"merge"} (default, saccades are extracted for each eye
independently but saccades from different eyes that temporally overlap are averaged into a binocular
saccade). Note that \code{binocular = "merge"} is overridden by \code{normalize = FALSE},
votes are left as they are per method and eye.}

\item{vote_threshold}{Value between 0..1 defining a vote threshold for a saccade.}
}
\value{
\code{data.frame} MORE INFO
}
\description{
Extract saccades from samples using votes from selected methods. Each method votes whether
a given sample belongs to a saccade. Next, saccades are identified via a majority vote using the 
\code{vote_threshold} parameter, as well as a minimal duration and minimal temporal separation criteria.

By default, ensemble includes methods proposed by Engbret & Kliegl (2003) (\code{"ek"}),
Otero-Millan et al. (\code{"om"}), and Nyström & Holmqvist (2010) (\code{"nh"}). However, it can be extended
via custom methods, see "Implementing custom methods" vignette.

By default, the function returns a table with identified saccades but can return a matrix with method's votes
per sample instead (\code{return_votes = TRUE}).
}
\details{
DETAILS ON RETURN SACCADE PROPERTIES
}
\examples{
data(single_trial)
ms <- extract_saccades(single_trial$x, single_trial$y, 500)
}
