% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_nh.R
\name{filter_via_savitzky_golay}
\alias{filter_via_savitzky_golay}
\title{Smooths signal using Savitzky-Golay and then shifts the filtered signal back}
\usage{
filter_via_savitzky_golay(x, sg_order)
}
\arguments{
\item{x}{}

\item{sgOrder}{}
}
\description{
Smooths signal using Savitzky-Golay and then shifts the filtered signal back
}
\keyword{internal}
