% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_on_samples.R
\name{vote_on_samples}
\alias{vote_on_samples}
\title{Computes ensemble vote for each sample using selected methods.}
\usage{
vote_on_samples(
  x,
  y,
  sample_rate,
  velocity_time_window = 20,
  methods = list("ek", "om", "nh"),
  binocular = "merge",
  normalize = TRUE,
  trial = NULL,
  options = NULL
)
}
\arguments{
\item{x}{Horizontal coordinate, ether a vector for monocular data or a two-column matrix for binocular data.}

\item{y}{Vertical coordinate, ether a vector for monocular data or a two-column matrix for binocular data.}

\item{sample_rate}{Sampling rate in Hz.}

\item{methods}{A _list_ (not a vector!) with names of package methods (character) or external functions that
implement sample classification (see vignette on using custom method). Package methods include 
Engbret & Kliegl (2003) (\code{"ek"}).}

\item{binocular}{Specifies how a binocular data is treated. Options are \code{"cyclopean"} (binocular data is
converted to an average cyclopean image before microsaccades are extracted), \code{"monocular"} (microsaccades
are extracted independently for each eye), \code{"merge"} (default, microsaccades are extracted for each eye
independently but microsaccades from different eyes that temporally overlap are averaged into a binocular
microsaccade). Note that \code{binocular = "merge"} is overridden by \code{normalize = FALSE},
votes are left as they are per method and eye.}

\item{normalize}{Logical, whether to aggregate and normalize votes to 0..1 range. Note that
\code{normalize = FALSE} overrides \code{binocular = "merge"}, votes are left as they are per method and eye.}

\item{trial}{Optional vector with trial ID. If omitted, all samples are assumed to belong to a single trial.}

\item{options}{A named list with options for a specific method.}
}
\value{
list Either a list of matrices (one per eye) with votes for each method (\code{normalized = FALSE}) or
a matrix with normalized (0..1) vote on each sample either for each eye or for single cyclopean or merged
binocular time series (\code{normalized = TRUE}).
}
\description{
This function is called by \code{\link{extract_microsaccades}} but can be used directly,
if you are interested in the vote-per-sample.
}
